\name{imguR-package}
\alias{imguR-package}
\docType{package}
\title{Imgur API Client Package for R}
\description{This package enables intuitive uploading of plots created in R to the free image hosting service [Imgur](http://www.imgur.com) simply, quickly, and intuitively via the [Imgur v3 API](http://api.imgur.com). The package is a complete client library for the Imgur API, meaning that one can also delete, favorite, and comment on images, as well as organize them in albums and galleries. By default, all images are loaded to Imgur anonymously. Optionally, using an OAuth2.0 login, users can also gain fully access to their own Imgur user account.}
\author{
Thomas J. Leeper and Aaron Statham

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://imgur.com/}{Imgur}

\href{https://api.imgur.com/}{API Documentation}
}
\keyword{ package }
%\seealso{}
%\examples{}
