\name{report_gallery_image}
\alias{report_gallery_image}
\title{Report a gallery image}
\description{Report a gallery image as inappropriate.}
\usage{
report_gallery_image(id, ...)
}
\arguments{
  \item{id}{An image ID.}
  \item{...}{Other arguments passed to HTTP request functions, for example: \code{token} (an OAuth2.0 token) or \code{key} (an API key, used by default).}
}
\details{Reports a gallery image as inappropriate.}
\value{An object of class \dQuote{imgur_basic}.}
%\references{}
\author{Thomas J. Leeper}
%\seealso{}
%\examples{}
